Ajout des constructeurs et des methodes getPosition

+++ Cosmos/src/cosmos/Position.java
@@ -10,10 +10,10 @@ package cosmos;
  * @author Marie Cogis David Goncalves
  */
 public class Position {
-    int x;
-    int y;
+    double x;
+    double y;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
